+++ R/ppareto.R
@@ -4,8 +4,8 @@
 #'  with shape parameter \eqn{\alpha}{alpha} and rate parameter
 #'  \eqn{\lambda}{lambda}.
 #'
-#' @template lowerthan-template
 #' @template q-template
+#' @template lower.tail-template
 #' @template pareto-template
 #'
 #' @export
@@ -13,15 +13,15 @@
 #' @examples
 #'
 #' # With scale parameter
-#' p_pareto(q = 2, shape = 2, scale = 5)
+#' ppareto(q = 2, shape = 2, scale = 5)
 #'
 #' # With rate parameter
-#' p_pareto(q = 2, shape = 2, rate = 5)
+#' ppareto(q = 2, shape = 2, rate = 5)
 #'
 #' # Survival function
-#' p_pareto(q = 2, shape = 2, rate = 5, lower.tail = FALSE)
+#' ppareto(q = 2, shape = 2, rate = 5, lower.tail = FALSE)
 #'
-p_pareto <- function(q, shape, rate = 1 / scale, scale = 1 / rate, lower.tail = TRUE) {
+ppareto <- function(q, shape, rate = 1 / scale, scale = 1 / rate, lower.tail = TRUE) {
     stopifnot(q >= 0, shape > 1, rate > 0)
 
     Sx <- (rate / (rate + q))^shape